Renamed hawfinch => swallow
