+++ .github/CONTRIBUTING.md
@@ -3,4 +3,4 @@
 Contributions are welcome, and they are greatly appreciated! Every little bit helps, and credit will always be given.
 
 Please read the Birdhouse [Contributer Guide](http://birdhouse.readthedocs.io/en/latest/contributing.html)
-and the [Cookiecutter Documentation](http://hawfinch.readthedocs.io/en/latest/) to get started.
+and the [Cookiecutter Documentation](http://swallow.readthedocs.io/en/latest/) to get started.
