+++ .github/ISSUE_TEMPLATE.md
@@ -4,7 +4,7 @@ Describe what you were trying to get done or your feature request.
 
 ## Environment
 
-* hawfinch version used, if any:
+* swallow version used, if any:
 * Python version, if any:
 * Operating System:
