+++ MANIFEST.in
@@ -1,6 +1,6 @@
 include Makefile
 include *.txt
 include *.rst
-recursive-include hawfinch *
+recursive-include swallow *
 global-exclude __pycache__
 global-exclude *.py[co]
