+++ docs/source/.gitignore
@@ -1,3 +1,3 @@
-/hawfinch.rst
-/hawfinch.*.rst
+/swallow.rst
+/swallow.*.rst
 /modules.rst
