+++ docs/source/conf.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-# hawfinch documentation build configuration file, created by
+# swallow documentation build configuration file, created by
 # sphinx-quickstart on Fri Jun  9 13:47:02 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -51,7 +51,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'hawfinch'
+project = u'swallow'
 copyright = u"2019, Tommy Godfrey"
 author = u"Tommy Godfrey"
 
@@ -113,7 +113,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'hawfinchdoc'
+htmlhelp_basename = 'swallowdoc'
 
 
 # -- Options for LaTeX output ------------------------------------------
@@ -140,8 +140,8 @@ latex_elements = {
 # (source start file, target name, title, author, documentclass
 # [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'hawfinch.tex',
-     u'hawfinch Documentation',
+    (master_doc, 'swallow.tex',
+     u'swallow Documentation',
      u'Tommy Godfrey', 'manual'),
 ]
 
@@ -151,8 +151,8 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'hawfinch',
-     u'hawfinch Documentation',
+    (master_doc, 'swallow',
+     u'swallow Documentation',
      [author], 1)
 ]
 
@@ -163,10 +163,10 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'hawfinch',
-     u'hawfinch Documentation',
+    (master_doc, 'swallow',
+     u'swallow Documentation',
      author,
-     'hawfinch',
+     'swallow',
      'One line description of project.',
      'Miscellaneous'),
 ]
