+++ environment.yml
@@ -1,4 +1,4 @@
-name: hawfinch
+name: swallow
 channels:
 - defaults
 - teriforey
