+++ swallow/__version__.py
