+++ swallow/default.cfg
@@ -1,10 +1,10 @@
 [metadata:main]
-identification_title = hawfinch
+identification_title = swallow
 identification_abstract = WPS to run the Met Office NAME model.
-identification_keywords = =PyWPS, WPS, OGC, processing, birdhouse, hawfinch, demo
+identification_keywords = =PyWPS, WPS, OGC, processing, birdhouse, swallow, demo
 identification_keywords_type = theme
-provider_name = hawfinch
-provider_url=http://hawfinch.readthedocs.org/en/latest/
+provider_name = swallow
+provider_url=http://swallow.readthedocs.org/en/latest/
 
 [server]
 url = http://0.0.0.0:5000/wps
@@ -16,5 +16,5 @@ parallelprocesses = 2
 
 [logging]
 level = INFO
-file = hawfinch.log
+file = swallow.log
 format = %(asctime)s] [%(levelname)s] line=%(lineno)s module=%(module)s %(message)s
