+++ swallow/swallow.py
