+++ swallow/utils.py
