+++ swallow/write_inputfile.py
