+++ swallow/write_scriptfile.py
