+++ tests/test_wps_name_standard.py
@@ -4,7 +4,7 @@ from pywps import Service
 from pywps.tests import assert_response_success
 
 from tests.common import client_for
-from hawfinch.processes.wps_name_standard import RunNAMEstandard
+from swallow.processes.wps_name_standard import RunNAMEstandard
 
 
 @pytest.mark.online
