+++ tests/test_wps_nameparams.py
@@ -4,7 +4,7 @@ from pywps import Service
 from pywps.tests import assert_response_success
 
 from tests.common import client_for
-from hawfinch.processes.wps_nameparams import RunNAME
+from swallow.processes.wps_nameparams import RunNAME
 
 
 @pytest.mark.skip
